pyproject.toml
[project]
name = "membery"
version = "0.1.0"
description = "A nostalgic memory matching card game"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "vtracer>=0.6.12",
]
